pyproject.toml
[project]
name = "demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncio>=3.4.3",
    "dotenv>=0.9.9",
    "fastapi>=0.115.12",
    "fastembed>=0.6.1",
    "langchain-groq>=0.3.2",
    "langchain-openai>=0.3.14",
    "mcp-use>=1.2.8",
    "mcp[cli]>=1.6.0",
    "openai-whisper>=20240930",
    "python-multipart>=0.0.20",
    "uvicorn>=0.34.2",
]
